Move create_pair_list to pairlistmanager
