--- freqtrade/data/common.py
@@ -1,7 +1,0 @@
-"""
-Common datatypes
-"""
-from typing import List, Tuple
-
-# List of pairs with their timeframes
-ListPairsWithTimeframes = List[Tuple[str, str]]
